Remove old soap scum logic
