+++ spec/soap_scum/ws_security_spec.rb
@@ -14,6 +14,16 @@ describe SoapScum::WSSecurity do
   let(:server_java_keyfile) { fixture_path('test_server.jks') }
   let(:keypass) { "importkey" }
 
+  let(:content_xml) do
+    <<-XML
+      <v4:listDocuments>
+        <v4:fileNumber>784449089</v4:fileNumber>
+      </v4:listDocuments>
+    XML
+  end
+  let(:content_document) { Nokogiri::XML(content_xml) }
+  let(:soap_document) { wrap_in_soap(content_document) }
+
   before do
     SoapScum::WSSecurity.configure(
       client_keyfile: client_keyfile,
@@ -22,26 +32,12 @@ describe SoapScum::WSSecurity do
     )
   end
 
-  describe '#encrypt' do
-    let(:content_xml) do
-      <<-XML
-        <v4:listDocuments>
-          <v4:fileNumber>784449089</v4:fileNumber>
-        </v4:listDocuments>
-      XML
-    end
-
+ describe '#encrypt' do
     let(:signed_elements) do
       [['/soapenv:Envelope/soapenv:Body', { soapenv: SoapScum::XMLNamespaces::SOAPENV }, 'Content']]
     end
 
-    let(:content_document) { Nokogiri::XML(content_xml) }
-    let(:soap_document) { wrap_in_soap(content_document) }
-    let(:java_encrypted_xml) { encrypted_xml_buffer(soap_document, 'listDocuments') }
-    let(:java_encrypted_doc) { parse_strict(java_encrypted_doc) }
-
     let(:soap_schema) { Nokogiri::XML::Schema(fixture('soap.xsd')) }
-
     let(:result) { SoapScum::WSSecurity.encrypt(soap_document, signed_elements) }
 
     it "returns valid SOAP" do
@@ -64,6 +60,17 @@ describe SoapScum::WSSecurity do
       )
 
       decrypted_doc = SoapScum::WSSecurity.decrypt(result.to_xml)
+      decrypted_doc = Nokogiri::XML(decrypted_doc)
+
+      expect(decrypted_doc.xpath('//v4:fileNumber').text).to eq('784449089')
+    end
+  end
+
+  describe '#decrypt' do
+    let(:java_encrypted_xml) { encrypted_xml_buffer(soap_document, fixture_path('test_server.jks'), 'listDocuments') }
+
+    it "it decrypts soap message encrypted by Java WSSecurity library" do
+      decrypted_doc = SoapScum::WSSecurity.decrypt(java_encrypted_xml)
       decrypted_doc = Nokogiri::XML(decrypted_doc)
 
       expect(decrypted_doc.xpath('//v4:fileNumber').text).to eq('784449089')
